rename config-files
