+++ shepherd_core/shepherd_core/data_models/content/virtual_harvester_config.py
